+++ src/refresh/vkpt/shader/global_ubo.h
@@ -163,9 +163,14 @@ with this program; if not, write to the Free Software Foundation, Inc.,
 	GLOBAL_UBO_VAR_LIST_DO(float,           god_rays_eccentricity) \
 	\
 	GLOBAL_UBO_VAR_LIST_DO(int,             num_cameras) \
-	GLOBAL_UBO_VAR_LIST_DO(int,             padding1) \
-	GLOBAL_UBO_VAR_LIST_DO(int,             padding2) \
-	GLOBAL_UBO_VAR_LIST_DO(int,             padding3) \
+	GLOBAL_UBO_VAR_LIST_DO(int,             screen_image_width) \
+	GLOBAL_UBO_VAR_LIST_DO(int,             screen_image_height) \
+	GLOBAL_UBO_VAR_LIST_DO(int,             prev_gpu_slice_width) \
+	\
+	GLOBAL_UBO_VAR_LIST_DO(int,             prev_width) \
+	GLOBAL_UBO_VAR_LIST_DO(int,             prev_height)\
+	GLOBAL_UBO_VAR_LIST_DO(float,           inv_width)\
+	GLOBAL_UBO_VAR_LIST_DO(float,           inv_height)\
 	\
 	GLOBAL_UBO_VAR_LIST_DO(vec4,            world_center) \
 	GLOBAL_UBO_VAR_LIST_DO(vec4,            world_size) \